pyproject.toml
[tool.poetry]
name = "schema-library"
version = "0.2.0"
description = ""
authors = ["OpsMill <info@opsmill.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10, < 3.13"
infrahub-sdk = {version = "^1.1.0", extras = ["all"]}
invoke = "^2.1.2"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
